Fixes #22
A PathResolver method that extracts the handler and format was broken
since the re-arrangement of the extensions. It was no longer properly
extracting the format of the template. The symptom was the layout file
was no longer being looked up properly (#22).

The sequence is as follows:
 - look up the template users/index
 - take the template and extract the handler and format, the format is
seen as null
 - set the template format as the new format to use
 - lookup the layout file with the new format to match the template
 - no layout file is found

The fix is to monkey patch the
`ActionView::PathResolver#extract_handler_and_format` method. It will
now make an additional attempt to lookup the types with another part of
the extension if it exists.

It is not ideal to monkey patch Rails methods, but in this case I think
it is necessary. The `extract_handler_and_format` method uses knowledge
of the constants set about the ‘registered details’ to extract the
extension data and since we’ve modified the registered details, we need
to modify the function as well.