Calling any helper method before set_version fails
The assumption in controller_additions is that the set_version before
filter will be called before anything starts using the helpers, and
before render is called. Because we are just a regular filter, another
controller calling prepend_before_filter can get in front of the setup.
If that filter were to reference any of the helpers, then they will be
nil.  It’s reasonable for this data to be available, because when the
filters are running in Rails the request is present with all the info
to extract the version.

This fix will allow the versioned_request to be instantiated at time of
any of the helper calls or if they are not, to instantiate in the
normal before filter.

Fixes #32