Allow configuration `missing_version` to be set to `unversioned_template`
`missing_version` now configures two options, one for `default_version`
and one for `missing_version_use_unversioned_template`.
`default_version` is used in the Rack layer to apply a version for a
missing version. `missing_version_use_unversioned_template` is used at
the rails layer to not fail when the request is missing a version and
no default version is set.